+++ notifications/urls.py
@@ -6,7 +6,7 @@ from . import views
 
 app_name = "notifications"
 urlpatterns = [
-    path("", views.AllNotificationsList.as_view(), name="all"),
+    path("all/", views.AllNotificationsList.as_view(), name="all"),
     path("unread/", views.UnreadNotificationsList.as_view(), name="unread"),
     path("mark-all-as-read/", views.mark_all_as_read, name="mark_all_as_read"),
     path("mark-as-read/<int:slug>)/", views.mark_as_read, name="mark_as_read"),