+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 1.8.3
+
+  - Fixes missing static folder/files
+
 ## 1.8.2
 
   - Added the migration for verbose_name changes